Move seed to an mjs file / ESM module
By declaring it as a ESM module we can avoid the workarounds when dealing with ESM-only dependencies

However, because we import the scenarios which are not declared as CJS we still get the [MODULE_TYPELESS_PACKAGE_JSON] warning that Node emits when importing ESM files from CJS and vice versa.
